move Next.js data fetching functions to /pages
