--- packages/app-content-pages/src/screens/Publications/getStaticProps.js
@@ -1,17 +1,0 @@
-import PublicationsAPI  from '../../api/publications'
-import { logNodeError } from '../../helpers/logger'
-
-export default async function getStaticProps() {
-  try {
-    const  publicationsData = await PublicationsAPI.createPublicationsResponse()
-    return {
-      props: {
-        publicationsData
-      },
-      revalidate: 60 * 60 * 1
-    }
-  } catch (error) {
-    logNodeError(error)
-    throw error
-  }
-}
