+++ packages/app-content-pages/src/screens/Publications/index.js
@@ -1,3 +1,1 @@
 export { default } from './PublicationsContainer'
-export { default as getStaticProps } from './getStaticProps'
-
