--- packages/app-content-pages/src/screens/Team/getStaticProps.js
@@ -1,17 +1,0 @@
-import TeamAPI  from '../../api/team'
-import { logNodeError } from '../../helpers/logger'
-
-export default async function getStaticProps() {
-  try {
-    const  teamData = await TeamAPI.createTeamResponse()
-    return {
-      props: {
-        teamData
-      },
-      revalidate: 60 * 60 * 1
-    }
-  } catch (error) {
-    logNodeError(error)
-    throw error
-  }
-}
